move visitCreateView to SparkSqlAstBuilder
