+++ sql/core/src/main/scala/org/apache/spark/sql/execution/SparkSqlParser.scala
@@ -27,7 +27,7 @@ import org.antlr.v4.runtime.{ParserRuleContext, Token}
 import org.antlr.v4.runtime.tree.TerminalNode
 
 import org.apache.spark.sql.catalyst.TableIdentifier
-import org.apache.spark.sql.catalyst.analysis.ViewType
+import org.apache.spark.sql.catalyst.analysis.{GlobalTempView, LocalTempView, PersistedView, UnresolvedDBObjectName}
 import org.apache.spark.sql.catalyst.catalog._
 import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.catalyst.parser._
@@ -413,26 +413,75 @@ class SparkSqlAstBuilder extends AstBuilder {
     }
   }
 
-  override def visitCreateTempView(
-      nameParts: Seq[String],
-      userSpecifiedColumns: Seq[(String, Option[String])],
-      comment: Option[String],
-      properties: Map[String, String],
-      originalText: Option[String],
-      child: LogicalPlan,
-      allowExisting: Boolean,
-      replace: Boolean,
-      viewType: ViewType): LogicalPlan = {
-    CreateViewCommand(
-      nameParts.asTableIdentifier,
-      userSpecifiedColumns,
-      comment,
-      properties,
-      originalText,
-      child,
-      allowExisting,
-      replace,
-      viewType = viewType)
+
+  /**
+   * Create or replace a view. This creates a [[CreateViewCommand]].
+   *
+   * For example:
+   * {{{
+   *   CREATE [OR REPLACE] [[GLOBAL] TEMPORARY] VIEW [IF NOT EXISTS] multi_part_name
+   *   [(column_name [COMMENT column_comment], ...) ]
+   *   create_view_clauses
+   *
+   *   AS SELECT ...;
+   *
+   *   create_view_clauses (order insensitive):
+   *     [COMMENT view_comment]
+   *     [TBLPROPERTIES (property_name = property_value, ...)]
+   * }}}
+   */
+  override def visitCreateView(ctx: CreateViewContext): LogicalPlan = withOrigin(ctx) {
+    if (!ctx.identifierList.isEmpty) {
+      operationNotAllowed("CREATE VIEW ... PARTITIONED ON", ctx)
+    }
+
+    checkDuplicateClauses(ctx.commentSpec(), "COMMENT", ctx)
+    checkDuplicateClauses(ctx.PARTITIONED, "PARTITIONED ON", ctx)
+    checkDuplicateClauses(ctx.TBLPROPERTIES, "TBLPROPERTIES", ctx)
+
+    val userSpecifiedColumns = Option(ctx.identifierCommentList).toSeq.flatMap { icl =>
+      icl.identifierComment.asScala.map { ic =>
+        ic.identifier.getText -> Option(ic.commentSpec()).map(visitCommentSpec)
+      }
+    }
+
+    val properties = ctx.tablePropertyList.asScala.headOption.map(visitPropertyKeyValues)
+      .getOrElse(Map.empty)
+    if (ctx.TEMPORARY != null && !properties.isEmpty) {
+      operationNotAllowed("TBLPROPERTIES can't coexist with CREATE TEMPORARY VIEW", ctx)
+    }
+
+    val viewType = if (ctx.TEMPORARY == null) {
+      PersistedView
+    } else if (ctx.GLOBAL != null) {
+      GlobalTempView
+    } else {
+      LocalTempView
+    }
+    if (viewType == PersistedView) {
+      CreateView(
+        UnresolvedDBObjectName(
+          visitMultipartIdentifier(ctx.multipartIdentifier),
+          false),
+        userSpecifiedColumns,
+        visitCommentSpecList(ctx.commentSpec()),
+        properties,
+        Option(source(ctx.query)),
+        plan(ctx.query),
+        ctx.EXISTS != null,
+        ctx.REPLACE != null)
+    } else {
+      CreateViewCommand(
+        visitMultipartIdentifier(ctx.multipartIdentifier).asTableIdentifier,
+        userSpecifiedColumns,
+        visitCommentSpecList(ctx.commentSpec()),
+        properties,
+        Option(source(ctx.query)),
+        plan(ctx.query),
+        ctx.EXISTS != null,
+        ctx.REPLACE != null,
+        viewType = viewType)
+    }
   }
 
   private def toStorageFormat(
